Added theme to the github repo since there will be changes made to the theme. Added values partial added values data Added css for values partial
Added committee partial - this uses the template for the services, which
is modified
Added committee data

Modified wrapper partial and params to act as a general about us partial

Removed some params for the contact form. gmaps location is uneccessary
along with phone and address.

Added files for the main pages of the site.

Fixed an issue with the logo being cut off on the homepage due to the
slider image. It now overlaps very nicely :)

I did all these changes before breakfast so there's probably more that I
forgot ~.~